Increment :attacks-performed-this-turn Fix #3 Fix #4
Regarding changes to tests. The fact that #3 and #4 had to be
reopened because some functionality had not been implemented
suggests that the tests were also insufficient. They now test that
`:attacks-performed-this-turn` is incremented. I have also changed
the style of the tests so that they compare the entire state
instead of just some specific values.