Add functionality for Sneed's Old Shredder Fix #38
In order to do this I also changed the mechanics for deathrattles
slightly. If the board if full before Sneed's dies a random
legendary minion should still be summoned. This means that the
deathrattle minion must be removed from the board before the
deathrattle effect is triggered. This is also consistent with the
order of animations in the game.